refactor(clean): estrarre config resolution in read_config.py
Separa le funzioni di sola configurazione dal runtime duckdb:
- _read_source_mode, _split_read_cfg, load_suggested_read,
  filter_suggested_read, resolve_clean_read_cfg → toolkit/clean/read_config.py
- duckdb_read.py ora importa da read_config (nessun contract esterno cambiato)
- backward compat: resolve_clean_read_cfg ancora raggiungibile da duckdb_read

Nessuna logica nuova, solo risistemazione dei confini tra moduli.
